requirements.txt
fastapi
pandas
torch
sentence-transformers
youtube-transcript-api
youtube-search
httpx
uvicorn
python-dotenv
cachetools
requests
hf_xet
Pillow